// src/app/models/propiedad.ts
export interface Propiedad {
  IdentificacionPropiedad: number;
  Referencia: string;
  TipoPropiedad: number;
  Operacion: number;
  MunicipioCod: number;
  Zona: number;
  Barrio: string;
  ConCartel: boolean;
  Valor: number;
  MetrosCuadrados: number;
  Dormitorios: number;
  Banos: number;
  Pisos: number;
  CobradorId: number;
  VendedorId: number;
  CodigoCliente: number;
  Telefono: string;
  FechaIngreso: string;
  Estado: number;
  PorcentajeAlquiler: number;
  Direccion: string;
  UsuarioModif: string;
}